feat: Add SourceRoot as a config option, defaulting to $GOPATH
The source root is the directory where the application is built and
prefixes callstack entries' file names. Removing this prefix improves
error stacktrace readability and grouping on the Bugsnag dashboard.

By default, the value is $GOPATH/src or $GOROOT/src if $GOPATH is not
set. At runtime, the value of $GOROOT is the root used during the Go
build.

Ref:

* https://golang.org/pkg/runtime/#GOROOT