fix(env): Ensure mock server shutdown
When the mock server thread is killed, there is a opportunity for the
server port to continue to be bound, which breaks future runs. This
change calls shutdown on the server if/when the thread is killed, and
failing that, prints in future runs that the mock server is not running.