[feat] Allow notifier.Notify to take any arguments
In order to idiomatically allow developers to pass in context.Contexts
as the first argument we need to change bugsnag.Notify to accept any
number of arguments, and then parse their types to get the errors,
context.Context, etc. required.

Benchmark indicated a negligible performance impact of about 400ns/op,
in particular compared to the performance gained in the last commit.