[fix] Prefer gofrs UUID library to satori's
satori/go.uuid introduced a breaking change a while back, which
can cause transient dependencies to be difficult to manage.

gofrs is an organisation of community members who work to ensure
the packages people depend on don't cause backwards breaking changes and
is therefore probably a better choice.

For more information:
https://github.com/satori/go.uuid/issues/66