[fix] Ensure fatal panics are notified
Due to a bug in the existing `notifier.NotifySync` method, we end up
using the notifier's config value to determine whether to send
synchronously or async. In the case of a fatal panic (i.e not
`AutoNotified` or `Recover`-ed) it still got sent asynchronously.
This change ensures that it does get sent synchronously.