[feat] support context request info for revel
Generally attempt to extract request information as late as possible
(i.e. just before sending of the report) but for Revel we have to
work around this a bit as Revel doesn't use `http.Request`s.
Instead of introducing revel dependencies (e.g. Revel's answer to
`http.Request`) in the main `bugsnag` package we extract the information
we can upfront upon receiving a request. Seeing as this is primarily
header information I don't foresee this being a big performance issue,
but there is a slightly larger overhead with Revel than with
`http.Request` based frameworks.