[fix] Make err the first parameter for Notify
This reverts a previous change that was made to make the go notifier a
bit more idiomatic. However, we don't intend on making the addition of
session tracking (the overarching work containing both the original
commit and this revert) a major release, semver-ily speaking.

Although, the change of signature from

```go
bugsnag.Notify(err error, rawData... interface{})
```

to

```go
bugsnag.Notify(rawData... interface{})
```

wouldn't break directly for users who just call this function as-is, it
would be breaking in the case that users decide to create a type alias
for the function signature of our client, in order to facilitate testing
or passing the function outside packages or libraries.