[fix] Ensure stackframes are skipped accurately
For session tracking we over-simplified the logic around when to skip
frames. This bug was not caught in our test suite. This commit fixes
these stackframe bugs, adds additional testing, and adds comments around
the locations where we skip stackframes to ensure this doesn't happen in
the future again.

Identified a bug in `(*Notifier).notify`, where the first frame would
always be from `(*Notifier).notify` itself. Skipping an additional frame
resolved this.

The test should highlight exactly what frames we expect for each of the
calls that would result in a report being sent. Apart from
`(*Notifier).Notify` these are consistent with the latest release.
That being said, whenever we're reporting panics (i.e. AutoNotify,
Recover) seem to be intentionally stripping off 2 low-level stackframes
(in the Go runtime) but also keeping 2. I am a bit confused by this, but
keeping for consistency.