[chore] Revel improvements
Managed to find where Revel is hiding the proper (*http.Request) object.
This allows us to get rid of a lot of custom code that we're relying on
to work around things for Revel.

Removing the check for version 0.18.0, as this is the new lower bound
for the versions of Revel we support.