[fix] Make notify release stages consistent
In particular, ensure that if the release stage is empty, it *will* send
the report.

In all fairness the fact that this has to be done through creating a
manual `bugsnag.Configuration` struct instead of using
`bugsnag.Configure` means that this is unlikely to cause confusion.

This is now consistent with other notifiers and how we treat
notify release stages for sessions.