Fix error_test.go
The output of Stack() isn't stable across versions of Go and shouldn't be relied on.
That makes the tests fail against Go 1.10, 1.11, and 1.12 (possibly earlier too).

Instead, the stack frames in Error should be verified against the stack frames supplied by the runtime.
Although the stack frames may vary from runtime impl to impl and version to version, we are
comparing like for like so the tests shouldn't break.