Also show inlined functions in stack trace for v1
This is a modified version of commit c6d59e66f3630108d84dcce59aaae871e4eb2b30
by Daria Bialobrzeska:

    Start showing inlined functions in stack trace (#208)

    * Start showing inlined functions in stack trace
    * Fix notifier tests
    * Added change to CHANGELOG

I selected only the code changes in the errors packages and also applied it
on the version 1 of this package (which is still used by notary as of writing).

This fixes the backtrace generated when using go 1.24, thus fixing the
TestUnwrapPkgError failure.